v0.2 [ALPHA]
Still in testing! Everything should be finished, though. In the next push, I will be implementing all the enchant limiters and custom crafting recipes, as well as possibly writing the readme if I get bored. I want to say that I don't think there will be any more major issues in the code, but I am not sure. Next push might be release, or it might be me getting pushed into a pit of more work.

:3